re-organize render
